Update AddPlanView to use PlanFormView
